scripting_with_python/sending_emails_with_python/requirements.txt
python-dotenv==0.19.2
